+++ src/rt/marker.rs
@@ -36,3 +36,15 @@ pub struct CheckSupportsInstanceProperty<T: SupportsInstanceProperty>(T);
 )]
 pub trait SupportsStaticProperty {}
 pub struct CheckSupportsStaticProperty<T: SupportsStaticProperty>(T);
+
+#[cfg(all(feature = "std", target_arch = "wasm32", panic = "unwind"))]
+use core::panic::UnwindSafe;
+
+/// Marker trait for types that are UnwindSafe only when building with panic unwind
+pub trait MaybeUnwindSafe {}
+
+#[cfg(all(feature = "std", target_arch = "wasm32", panic = "unwind"))]
+impl<T: UnwindSafe + ?Sized> MaybeUnwindSafe for T {}
+
+#[cfg(not(all(feature = "std", target_arch = "wasm32", panic = "unwind")))]
+impl<T: ?Sized> MaybeUnwindSafe for T {}