+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 
 ### Added
 
+* Support catching panics, and raising JS Exceptions for them, when building
+  with panic=unwind on nightly, with the `std` feature.
+  [#4790](https://github.com/wasm-bindgen/wasm-bindgen/pull/4790)
+
 * Added support for passing `&[JsValue]` slices from Rust to JavaScript functions.
   [#4872](https://github.com/wasm-bindgen/wasm-bindgen/pull/4872)